+++ example/modules/Items.ts
@@ -1,12 +1,16 @@
+import { Modules } from '.'
 import { StonexModule } from '../../src'
 
-export default class Items extends StonexModule<{ data: number[]; isLoading: boolean }> {
+export default class Items extends StonexModule<{ data: number[]; isLoading: boolean }, Modules> {
   public state = {
     data: [],
     isLoading: false
   }
 
   public getList = async () => {
+
+    console.log('get list called')
+
     if (this.state.data.length) {
       return this.state.data
     }
@@ -14,6 +18,7 @@ export default class Items extends StonexModule<{ data: number[]; isLoading: boo
     this.setState({ isLoading: true })
     const data = await Promise.resolve([1, 2, 3, 4, 5, 6])
     this.setState({ data })
+
     return this.state.data
   }
 }