+++ CHANGELOG.md
@@ -12,6 +12,15 @@
  -->
 
 
+## [0.1.3-beta](https://github.com/acacode/stonex/releases/tag/0.1.3-beta)
+
+### Added
+- The bridge for using another modules in module via using property `this.modules` inside methods of your StonexModule
+
+### Changed
+- Files of example of stonex store. Tested new feature inside code example (folder: `example`)
+
+
 ## [0.1.3-alpha](https://github.com/acacode/stonex/releases/tag/0.1.3-alpha)
 
 ### Changed